requirements.txt
selenium==3.141.0
selenium-requests==1.3.3
text2emotion==0.0.5
pyfiglet==0.8.post1
matplotlib==3.5.0
chromedriver-autoinstaller-0.2.2
